feat: add desciption for VIP 343
